MoCoRP/requirements.txt
boto3==1.9.246
botocore==1.12.246
docutils==0.14
emoji==0.5.4
docformatter==1.3.0
flake8==3.7.8
flake8-bugbear==19.8.0
GitPython==3.0.3
h5py==2.10.0
joblib>=0.14.0
nltk==3.4.5
numpy>=1.16.0
pexpect==4.7.0
Pillow>=6.2.0
py-gfm==0.1.4
py-rouge==1.1
pyzmq==18.1.0
regex==2019.8.19
recommonmark==0.6.0
requests==2.22.0
requests-mock==1.7.0
scikit-learn>=0.21.0
scipy>=1.3.0
sh==1.12.14
Sphinx==2.2.0
sphinx_rtd_theme==0.4.3
sphinx-autodoc-typehints==1.10.3
Unidecode==1.1.1
websocket-client==0.56.0
websocket-server==0.4
torch==1.12.0
transformers==4.36.1
jsonlines==4.0.0
wandb==0.17.0